requirements.txt
# Core App
streamlit==1.32.0
boto3==1.34.67

# PDF Export
fpdf==1.7.2

# Optional (for formatting or deployment)
python-dotenv==1.0.1

# If you ever use dataframes or further analysis
pandas==2.2.1
